backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trustagency-backend"
version = "1.0.0"
description = "Admin CMS with AI Content Generation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "TrustAgency Team", email = "dev@trustagency.com"}
]
keywords = ["fastapi", "admin", "cms", "ai", "content-generation"]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "psycopg2-binary==2.9.9",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "PyJWT==2.8.1",
    "fastapi-admin==0.3.3",
    "celery==5.3.4",
    "redis==5.0.1",
    "openai==1.3.5",
    "requests==2.31.0",
    "python-slugify==8.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
